+++ lib/modules/generals/newtypes/domain.d.ts
