+++ lib/modules/generals/newtypes/infrastructure.d.ts
